requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.4.5.1
chardet==3.0.4
coloredlogs==14.0
decorator==4.4.2
flake8==3.8.2
humanfriendly==8.2
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
jupyter-client==6.1.5
jupyter-core==4.6.3
mccabe==0.6.1
numpy==1.18.5
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
